Move save recipe route and tests to actions module
