+++ meal_planner/form_processing.py
@@ -1,0 +1,23 @@
+from starlette.datastructures import FormData
+
+
+def _parse_recipe_form_data(form_data: FormData, prefix: str = "") -> dict:
+    """Parses recipe form data into a dictionary, handling optional prefix."""
+    name_value = form_data.get(f"{prefix}name")
+    name = name_value if isinstance(name_value, str) else ""
+
+    ingredients_values = form_data.getlist(f"{prefix}ingredients")
+    ingredients = [
+        ing for ing in ingredients_values if isinstance(ing, str) and ing.strip()
+    ]
+
+    instructions_values = form_data.getlist(f"{prefix}instructions")
+    instructions = [
+        inst for inst in instructions_values if isinstance(inst, str) and inst.strip()
+    ]
+
+    return {
+        "name": name,
+        "ingredients": ingredients,
+        "instructions": instructions,
+    }
